Cuatro fallos que el founder vio hoy y yo no
1. /auth/validate-invite devolvia 500 en produccion. El modelo declaraba
   tokens:int y el plan Medio da 5 + 2160/100 = 26,6 — Pydantic lo rechazaba.
   El formulario de registro se quedaba MUDO al teclear PROINV-A04748, porque
   el 5xx cae en la rama que no pinta nada. Lo rompi yo ayer al arreglar el "?".

2. Ningun caso cognitivista se estaba guardando en el servidor. El frontend
   manda case_type='cognitiva' desde el 12-sep y el backend solo aceptaba
   behavior|training|puppy: 422, mres.ok false, todo dentro de un try/catch
   mudo, y el boton decia "Caso salvato" igual. Vivian solo en el movil, sin
   ejercicios diarios y sin PDF. Ahora se acepta, y si el servidor no lo coge,
   se dice.

3. Abrir un caso cognitivista guardado enseñaba un analisis ABA de OTRO perro.
   openRecord deducia la postura mirando solo `analysis`, que en cognitivista
   va vacio a proposito (el texto esta en `plan`) -> 'behavioral' -> pantalla
   ABC, y como no habia nada con que repintarla quedaba a la vista el caso
   anterior. Ahora la postura sale de stance, de case_type y del lexico de los
   dos campos, y un caso cognitivista abre SU relazione guardada, tal cual se
   guardo. Comprobado en navegador: cae en s-relazione-cz, no en s-abc, y el
   conductual sigue yendo a s-abc.

4. La via cognitivista era la unica sin PDF: no existia ni boton ni endpoint.
   Nuevo GET /cases/{id}/cognitiva/pdf + build_cognitiva_pdf, rotulos en
   italiano y sin vocabulario conductual, y boton "Scarica il PDF".

Y dos encargos suyos del mismo rato:
 · Fuera la derivacion a profesional. La relazione organizaba el plan en
   incontri "guidato da un professionista" y Escuela de Cachorros tenia un
   CIERRE OBLIGATORIO que derivaba, en los tres idiomas. Sus palabras: si quien
   usa la app es profesional no tiene sentido mandarlo a un profesional, y si
   es un particular esta aqui para intentarlo el. Las prioridades MEDICAS se
   quedan: eso es medicina, no un rinvio.
 · Cecilia en cachorros y en adiestramiento, NO en cognitivista (canal estanco:
   ella explica en lenguaje conductual).

+++ app/api/routes/auth.py
@@ -105,7 +105,12 @@ class ValidateInviteResponse(BaseModel):
     valid: bool
     type: Optional[str] = None       # 'delegation' | 'ambassador' | None
     label: Optional[str] = None      # nombre legible ("Bocalán Chile", "Embajador")
-    tokens: Optional[int] = None     # total welcome tokens si valid
+    # 19-sep-2026: era `int` y reventaba con un 500. Los tokens del sistema son
+    # FRACCIONARIOS (1 tk = 100 creditos), asi que el plan Medio da 2160 cr =
+    # 21,6 tk, y 5 + 21,6 = 26,6 — que Pydantic rechaza contra un `int`. El
+    # formulario de registro se quedaba mudo al teclear PROINV-A04748: el 5xx
+    # cae en la rama que no pinta nada. Tipo correcto: float, como la columna.
+    tokens: Optional[float] = None   # total welcome tokens si valid
 
 
 # ── Helpers ───────────────────────────────────────────────────────────────────
